create separate dialog components, reduce some duplicated code
