+++ client/src/components/modal-forms/index.js
@@ -17,4 +17,4 @@ export * from './EditRosStartDateForm';
 export * from './EditRosSiteForm';
 export * from './PhaseDialog';
 export * from './BulkAllocationForm';
-export * from './UserManagementForm';
+export * from './UserManagementViewForm';
